Database updated

+++ app/src/main/java/com/example/nisalikularatne/runningtracker/data/DBHelper.java
@@ -8,13 +8,14 @@ import android.database.sqlite.SQLiteOpenHelper;
 import com.example.nisalikularatne.runningtracker.RunnerTracker;
 
 public class DBHelper extends SQLiteOpenHelper {
-    private static final int DATABASE_VERSION = 1;
+    private static final int DATABASE_VERSION = 5;
     private static final String DATABASE_NAME = "RunnerTrackerDB.db";
-    public static final String TABLE_RUNNERTRACKER = "RunnerTracker";
+    public static final String TABLE_RUNNERTRACKER = "RunnerTracker2";
 
     public static final String COLUMN_ID = "_id";
     public static final String COLUMN_DISTANCE = "distance";
     public static final String COLUMN_DATE = "date";
+    public static final String COLUMN_TIME = "time";
 
     public DBHelper(Context context, String name,
                        SQLiteDatabase.CursorFactory factory, int version) {
@@ -25,21 +26,26 @@ public class DBHelper extends SQLiteOpenHelper {
         String CREATE_RUNNERTRACKER_TABLE = "CREATE TABLE " +
                 TABLE_RUNNERTRACKER + "("
                 + COLUMN_ID + " INTEGER PRIMARY KEY," + COLUMN_DISTANCE
-                + " INTEGER," + COLUMN_DATE + " TEXT" + ")";
+                + " INTEGER," + COLUMN_DATE + " TEXT,"+
+        COLUMN_TIME + " TEXT"+ ")";
         db.execSQL(CREATE_RUNNERTRACKER_TABLE);
 
     }
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion,
                           int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_RUNNERTRACKER);
-        onCreate(db);
+        if (newVersion > oldVersion) {
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_RUNNERTRACKER);
+            onCreate(db);
+        }
+
     }
     public void addRunnerTracker(RunnerTracker runnerTracker) {
 
         ContentValues values = new ContentValues();
         values.put(COLUMN_DISTANCE, runnerTracker.getRunnerTrackerDistance());
         values.put(COLUMN_DATE, runnerTracker.getRunnerTrackerDate());
+        values.put(COLUMN_TIME, runnerTracker.getRunnerTrackerTime());
 
         SQLiteDatabase db = this.getWritableDatabase();
